requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
efficientnet-pytorch==0.7.1
frozenlist==1.3.3
fsspec==2022.11.0
idna==3.4
joblib==1.2.0
lightning-utilities==0.4.2
multidict==6.0.3
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.1
python-dateutil==2.8.2
pytorch-lightning==1.8.4.post0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.13.0
torchmetrics==0.11.0
torchvision==0.14.0
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
